+++ packages/gamut/src/Form/SelectDropdown/SelectDropdown.tsx
@@ -18,6 +18,7 @@ import {
   DropdownButton,
   formatGroupLabel,
   formatOptionLabel,
+  IconOption,
   MultiValueRemoveButton,
   MultiValueWithColorMode,
   onFocus,
@@ -44,7 +45,7 @@ const defaultProps = {
     SelectContainer: CustomContainer,
     MultiValue: MultiValueWithColorMode,
     MultiValueRemove: MultiValueRemoveButton,
-    Option: SelectDropdownOption,
+    Option: IconOption,
   },
 };