+++ packages/gamut/src/Layout/LayoutGrid.tsx
@@ -1,7 +1,7 @@
 import { pxRem, styledOptions, system, theme } from '@codecademy/gamut-styles';
 import { StyleProps, variance } from '@codecademy/variance';
 import styled from '@emotion/styled';
-import { pick } from 'lodash';
+import pick from 'lodash/pick';
 
 const columns = 12;