+++ packages/styleguide/.storybook/addons/system/components/PropsTable.tsx
@@ -2,7 +2,7 @@ import { ComponentProps, useMemo } from 'react';
 import * as React from 'react';
 import { PROP_GROUPS } from '../propMeta';
 import { ArgsTable } from '@storybook/addon-docs/blocks';
-import { isArray } from 'lodash';
+import isArray from 'lodash/isArray';
 
 import {
   PropItem,